Ya termine el tema de las tablas, solo faltaria la de mediciones del sensor. Hay que reveer el tema de como el cliente se asigna dispositivos ya que ahora lo que deberia hacer es crearse una instancia de DispositivoXCliente pasandole como parametros al cliente y al dispositivo a agregar, y por ultimo persistir el DispositivoXCliente

+++ src/main/java/modelsPersistencia/ClienteModel.java
@@ -12,7 +12,7 @@ public class ClienteModel extends ModelHelperPersistencia{
 		super();
 	}
 	
-	public Cliente buscarCliente(int id) {
+	public Cliente buscarCliente(Long id) {
 		return super.buscar(Cliente.class, new ImmutablePair<>("id", id));
 	}
